Add users to functionalities tables. Create forbidden exception. Beggining of creating service for resource authorization for connection functionality

+++ EnglishJourney.Application/Connection/Commands/DeleteAttribute/DeleteAttributeCommandHandler.cs
@@ -1,12 +1,15 @@
-﻿using EnglishJourney.Domain.Entities;
+﻿using EnglishJourney.Domain.Constants;
+using EnglishJourney.Domain.Entities;
 using EnglishJourney.Domain.Exceptions;
 using EnglishJourney.Domain.Interfaces;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 
 namespace EnglishJourney.Application.Connection.Commands.DeleteAttribute
 {
-    public class DeleteAttributeCommandHandler(IConnectionRepository repository, ILogger<DeleteAttributeCommandHandler> logger) : IRequestHandler<DeleteAttributeCommand>
+    public class DeleteAttributeCommandHandler(IConnectionRepository repository, ILogger<DeleteAttributeCommandHandler> logger,
+        IEnglishJourneyAuthorizationService authorizationService) : IRequestHandler<DeleteAttributeCommand>
     {
         public async Task Handle(DeleteAttributeCommand request, CancellationToken cancellationToken)
         {
@@ -14,6 +17,9 @@ namespace EnglishJourney.Application.Connection.Commands.DeleteAttribute
 
             var attribute = await repository.GetAttributeById(request.Id);
             if (attribute == null) throw new NotFoundException(nameof(ConnectionAttribute), request.Id.ToString());
+
+            if (!authorizationService.AuthorizeConnection(attribute.Topic, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await repository.DeleteAttribute(attribute);
         }
