+++ EnglishJourney.Application/Connection/Commands/DeleteTopic/DeleteTopicCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using EnglishJourney.Domain.Entities;
+﻿using EnglishJourney.Domain.Constants;
+using EnglishJourney.Domain.Entities;
 using EnglishJourney.Domain.Exceptions;
 using EnglishJourney.Domain.Interfaces;
 using MediatR;
@@ -6,7 +7,9 @@ using Microsoft.Extensions.Logging;
 
 namespace EnglishJourney.Application.Connection.Commands.DeleteTopic
 {
-    public class DeleteTopicCommandHandler(IConnectionRepository repository, ILogger<DeleteTopicCommandHandler> logger) : IRequestHandler<DeleteTopicCommand>
+    public class DeleteTopicCommandHandler(IConnectionRepository repository,
+        IEnglishJourneyAuthorizationService authorizationService,
+        ILogger<DeleteTopicCommandHandler> logger) : IRequestHandler<DeleteTopicCommand>
     {
         public async Task Handle(DeleteTopicCommand request, CancellationToken cancellationToken)
         {
@@ -14,6 +17,9 @@ namespace EnglishJourney.Application.Connection.Commands.DeleteTopic
 
             var topic = await repository.GetTopicById(request.Id);
             if (topic == null) throw new NotFoundException(nameof(ConnectionTopic), request.Id.ToString());
+
+            if (!authorizationService.AuthorizeConnection(topic, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await repository.DeleteTopic(topic);
         }
